feat: Added a way to lock viewport to a specific image control name (for events). chore: Update 'TODO.md' by moving the export game point to the preview 2. feat: Added the event on the "UI Editor" menu button so now it open the UI Editor. feat: Added the viewport to the UI Editor. Also tested the multi-viewport feature.

+++ RPGCreator.RTP/Viewport/MonogameViewport.cs
@@ -222,7 +222,7 @@ public class MonogameViewport : BaseMonogameViewport
 
         var address = _bitmapControlAddress.Value;
         
-        int totalPixels = (Size.Width * Size.Height);
+        int totalPixels = ((int)Size.Width * (int)Size.Height);
         
         if (_internalBuffer == null || _internalBuffer.Length != totalPixels)
         {
