+++ RPGCreator.RTP/Viewport/UiViewport.cs
@@ -122,7 +122,7 @@ public class UiViewport : BaseMonogameViewport
 
         var address = _bitmapControlAddress.Value;
         
-        int totalPixels = (Size.Width * Size.Height);
+        int totalPixels = ((int)Size.Width * (int)Size.Height);
         
         if (_internalBuffer == null || _internalBuffer.Length != totalPixels)
         {
